requirements.txt
pillow
torch
transformers
hydra-core
pandas
torchvision
omegaconf
clip-openai